+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 
 ## 0.1.0(2024/07/08)
 - initial release
+
+## 0.2.0(2024/07/08)
+- change export format from pickle to codecs.
+- add parameters to `add_glob` function.
+  - as_tree
+  - include_ext
+- add parameters to `export` function.
+  - create_pyc
+- add help text and type linting.